requirements.txt
# Core Dependencies
google-cloud-bigquery==3.13.0
google-cloud-functions==1.13.0
google-cloud-secret-manager==2.18.0
google-cloud-scheduler==2.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# API Clients
simple-salesforce==1.12.6
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
hubspot-api-client==7.0.0
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
phonenumbers==8.13.26

# LLM & AI
anthropic==0.7.8
openai==1.3.7
google-cloud-aiplatform==1.38.1

# Vector Search & Embeddings
sentence-transformers==2.2.2

# Web Framework (for Streamlit option)
streamlit==1.28.2
streamlit-authenticator==0.2.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging & Monitoring
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-pubsub==2.21.5

# Email Processing
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16

